+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.7] 2025-03-19
+
+### Changed
+- The slicer to choose time granularity in the results panel is now 
+  dedicated for the bar chart instead of the line chart.
+
 ## [0.6.6] 2025-03-20
 
 ### Changed